integration of hooks.c into main library
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rulebasedmodels@305 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/C5/src/formtree.c
@@ -119,7 +119,7 @@ void InitialiseTreeData()
     }
 
     ClassFreq = AllocZero(MaxClass+1, double);
-    ClassSum  = Alloc(MaxClass+1, float);
+    ClassSum  = Alloc(MaxClass+1, double);
 
     if ( BOOST )
     {
